+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 1.1.43 - 2026-06-29
 ### Fixed
 - Fixed Auto Restore Calibration option in phd2 guiding
+- Fixed issue where max slew rate required mount reconnect
 
 ## 1.1.42 - 2026-06-24
 ### Added