+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.43 - 2026-06-29
+### Fixed
+- Fixed Auto Restore Calibration option in phd2 guiding
+- Fixed issue where max slew rate required mount reconnect
+
 ## 1.1.42 - 2026-06-24
 ### Added
 - SequenceItem to slew to phd2 calibration position